refactor: util を utils に改名
